applications/neural_search/requirements.txt
pymilvus>=1.1.2
pandas==0.25.1 
paddlenlp>=2.1.1    
paddlepaddle-gpu>=2.1.3
hnswlib>=0.5.2
numpy>=1.17.2
visualdl>=2.2.2
paddle-serving-app>=0.7.0        
paddle-serving-client>=0.7.0        
paddle-serving-server-gpu>=0.7.0.post102
pybind11
